Compute the β (beta) parameter for the SiGN model

Computes the β adjustment for signalled alternatives, reflecting the balance between conditional and terminal (i.e., primary/unconditional) reinforcement, as described in Equation 6 of Dunn et al. (2024).

Usage

beta_sig(
  use_beta,
  sig,
  sched1,
  sched2,
  il_dur1,
  il_dur2,
  s_plus_dur,
  base = 10
)

Arguments

use_beta

Logical vector indicating whether the β parameter should be applied.

sig

Logical vector indicating whether the alternative is signalled. See sig_check.

sched1

Character vector indicating the schedule type for the first alternative's initial link. See Details. (e.g., "VI", "FR").

sched2

Character vector indicating the schedule type for the second alternative's initial link. See Details.

il_dur1

Numeric vector of durations for the first alternative's initial link.

il_dur2

Numeric vector of durations for the second alternative's initial link.

s_plus_dur

Numeric vector of durations associated with alternative 1's \(S^{+}\) stimulus.

base

A positive real number specifying the base of the logarithm used for β. Defaults to 10.

Value

A numeric vector containing the β values.

Details

Although the initial link durations for both alternatives are provided as arguments, the function computes β specifically for the first initial link.

If both initial links are variable interval (VI) schedules, a common initial link duration is computed: $$1 / (1 / d_1 + 1 / d_2)$$ Otherwise, the duration of the first initial link is used.

See choice_params() for additional details appropriate on schedule selection.

References

Dunn, R. M., Pisklak, J. M., McDevitt, M. A., & Spetch, M. L. (2024). Suboptimal choice: A review and quantification of the signal for good news (SiGN) model. Psychological Review. 131(1), 58-78. https://doi.org/10.1037/rev0000416

Examples

beta_sig(
  use_beta = TRUE,
  sig = TRUE,
  sched1 = "FR",
  sched2 = "FR",
  il_dur1 = 1,
  il_dur2 = 1,
  s_plus_dur = 10
)
#> [1] 1.041393