List Recycle

Recycle list elements to equal length.

Usage

recycle_list(x)

Arguments

x

A named or unnamed list where each element is a vector. Elements may vary in length.

Value

A list where each element has length equal to the maximum length of the original list elements. Shorter elements are repeated to match this length.

Examples

recycle_list(list(a = 1:3, b = 5))
#> $a
#> [1] 1 2 3
#> 
#> $b
#> [1] 5 5 5
#> 

recycle_list(list(a = 1:7, b = c("Pigeon", "Pigeon", "Rat")))
#> $a
#> [1] 1 2 3 4 5 6 7
#> 
#> $b
#> [1] "Pigeon" "Pigeon" "Rat"    "Pigeon" "Pigeon" "Rat"    "Pigeon"
#>